fix supporters

+++ project/search/controllers.js
@@ -1,7 +1,7 @@
 import {BaseLayoutController} from 'outlinejs/lib/controllers';
 import {runtime} from 'outlinejs/lib/contexts';
 import {MaterialLayoutView} from '../core/views';
-import {UserSkillCollection} from '../skills/managers';
+import {UserSkillCollection, SupportUserSkillCollection} from '../skills/managers';
 import {Skill, SupportUserSkill} from '../skills/models';
 import {SearchContentView} from './views';
 import {globalContext} from '../global';
@@ -28,17 +28,42 @@ export class SearchSkillContoller extends BaseLayoutController {
     };
   }
 
-  async unSupportUser(supportUserSkill) {
-    try {
-      let userFullName = supportUserSkill.userFullName;
-      let skillName = supportUserSkill.skillName;
-      await supportUserSkill.destroy({headers: {'content-type': 'application/json'}});
-      GuiNotifications.snackBar(`Your support for ${userFullName} in ${skillName} has been removed!`);
-      return true;
-    } catch (err) {
-      GuiNotifications.snackBar('Ops! Something went wrong!');
-      console.log(err);
-      return false;
+  async unSupportUser(supportUserSkill = null, userSkill = null, supporter = this.request.user.id) {
+    // priority to supportUserSkill, else find a supportUserSkill with supporter and userSkill and delete it
+    if (supportUserSkill !== null) {
+      try {
+        let userFullName = supportUserSkill.userFullName || userSkill.userFullName;
+        let skillName = supportUserSkill.skillName || userSkill.skillName;
+        await supportUserSkill.destroy({headers: {'content-type': 'application/json'}});
+        if (userSkill) {
+          userSkill.supporters = userSkill.supporters.filter(item => item !== supporter);
+        }
+        this.render(this.context);
+
+        GuiNotifications.snackBar(`Your support for ${userFullName} in ${skillName} has been removed!`);
+        return true;
+      } catch (err) {
+        GuiNotifications.snackBar('Ops! Something went wrong!');
+        console.log(err);
+        return false;
+      }
+    } else if (supporter && userSkill) {
+
+      try {
+        let supportUserSkills = new SupportUserSkillCollection();
+        await supportUserSkills.filterBySupporterAndUserSkill(supporter, userSkill.id);
+        if (supportUserSkills.length === 1) {
+          supportUserSkill = supportUserSkills.at(0);
+          this.unSupportUser(supportUserSkill, userSkill);
+        }
+
+      } catch (err) {
+        GuiNotifications.snackBar('Ops! Something went wrong!');
+        console.log(err);
+        return false;
+      }
+    } else {
+      console.log('Give to unSupportUser supportUserSkill or supporter and userSkill');
     }
   }
 
@@ -53,7 +78,7 @@ export class SearchSkillContoller extends BaseLayoutController {
       try {
         await supportUserSkill.save();
         userSkill.supporters.push(this.request.user.id);
-        GuiNotifications.snackBar(`Now you support ${userSkill.userFullName} in  ${userSkill.skillName}!`, 'Undo', this.unSupportUser.bind(this, supportUserSkill));
+        GuiNotifications.snackBar(`Now you support ${userSkill.userFullName} in  ${userSkill.skillName}!`, 'Undo', this.unSupportUser.bind(this, supportUserSkill, userSkill, this.request.user.id));
         this.render(this.context);
         return true;
       } catch (err) {
